Implement strafing.
Make A/D strafe left/right, and use j/l to turn.

This is a more traditional interpretation of WASD controls, and being
able to strafe around corners is a classic vital FPS technique.

Ideally we'd use the arrow keys for turning, but that'd require
revamping the input mechanism somewhat (probably doable via `stty
-cbreak` or so).